Made every Mathematical Function Container extend RecursiveValidatedService

+++ src/main/java/nl/smith/mathematics/numbertype/RationalNumber.java
@@ -21,7 +21,7 @@ public class RationalNumber extends Number implements ArithmeticFunctions<Ration
 
     private final BigInteger denominator;
 
-    public RationalNumber(int numerator) {
+    public RationalNumber(long numerator) {
         this(BigInteger.valueOf(numerator));
     }
